+++ src/main/java/pl/agh/edu/wiet/to2/kevin/service/questions/scoring/strategies/impl/FractionalWithSignStrategy.java
@@ -5,13 +5,16 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
 import pl.agh.edu.wiet.to2.kevin.service.questions.scoring.strategies.BaseScoringStrategy;
 
+import java.math.BigDecimal;
+
 @Service
 @Qualifier("fractionalWithSignScoringStrategy")
-@Scope("singleton") public final class FractionalWithSignStrategy extends BaseScoringStrategy {
+@Scope("singleton")
+public final class FractionalWithSignStrategy extends BaseScoringStrategy {
 
     private static final String STRATEGY_NAME = "Fractional point";
 
     public FractionalWithSignStrategy() {
-        super(false, true, true, 1, -1, STRATEGY_NAME);
+        super(false, true, true, BigDecimal.valueOf(1), BigDecimal.valueOf(-1), STRATEGY_NAME);
     }
 }
