+++ src/main/java/pl/agh/edu/wiet/to2/kevin/service/questions/scoring/strategies/impl/PointForPointStrategy.java
@@ -5,13 +5,16 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
 import pl.agh.edu.wiet.to2.kevin.service.questions.scoring.strategies.BaseScoringStrategy;
 
+import java.math.BigDecimal;
+
 @Service
 @Qualifier("pointForPointScoringStrategy")
-@Scope("singleton") public final class PointForPointStrategy extends BaseScoringStrategy {
+@Scope("singleton")
+public final class PointForPointStrategy extends BaseScoringStrategy {
 
     private static final String STRATEGY_NAME = "Point for point";
 
     public PointForPointStrategy() {
-        super(false, true, false, 1, -1, STRATEGY_NAME);
+        super(false, true, false, BigDecimal.valueOf(1), BigDecimal.valueOf(-1), STRATEGY_NAME);
     }
 }
