+++ src/main/java/pl/agh/edu/wiet/to2/kevin/service/questions/scoring/strategies/impl/ZeroOnErrorStrategy.java
@@ -5,13 +5,16 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
 import pl.agh.edu.wiet.to2.kevin.service.questions.scoring.strategies.BaseScoringStrategy;
 
+import java.math.BigDecimal;
+
 @Service
 @Qualifier("zeroOnErrorScoringStrategy")
-@Scope("singleton") public final class ZeroOnErrorStrategy extends BaseScoringStrategy {
+@Scope("singleton")
+public final class ZeroOnErrorStrategy extends BaseScoringStrategy {
 
     private static final String STRATEGY_NAME = "Zero on error";
 
     public ZeroOnErrorStrategy() {
-        super(true, false, false, 1, 0, STRATEGY_NAME);
+        super(true, false, false, BigDecimal.valueOf(1), BigDecimal.valueOf(0), STRATEGY_NAME);
     }
 }
